Initial github release --Version 1.0.3
Lots of code cleanup.
Added a Save button so you can get into the part where you type in the set name with 1 click.
Moved the Previous, Next, Back buttons into their own frame below the Gossip window. Mostly because i wanted to add the save button, but also because some times the left most buttons get a graphic glitch with the texture. I think something in the default frame is overlapping but nothing shows in the framestack.